Migrate constants
